Refactorizar para implementar nueva estructura de clases del modelo
